update release workflow
check out the master branch insted, as it wants to update the
README.md on the master branch. Not in the detached head of the
release.

set up git user/email to be clear it's from GitHub Actions, and
also because in some cases the action won't be able to infer the
author